# contracts/internal/internal-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SuperAgentSystem Internal API
  version: v1
  summary: Launch Spec v1 frozen internal contracts with Memory System V3 single-tenant extension
servers:
  - url: https://internal.super-agent.local
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
    Scope:
      name: X-Scope
      in: header
      required: true
      schema:
        type: string
    TraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Standard error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MemoryQueryRequest:
      type: object
      properties:
        kind:
          type: string
          enum: [resident, factual, procedural, summary, candidate]
        task_request_id:
          type: string
          format: uuid
        fingerprint:
          type: string
        limit:
          type: integer
      additionalProperties: false
    MemoryQueryResponse:
      type: object
      required: [kind, tenant_id, scope, single_tenant_mode, items]
      properties:
        kind:
          type: string
        tenant_id:
          type: string
        scope:
          type: string
        single_tenant_mode:
          type: boolean
        items:
          type: array
          items:
            type: object
            additionalProperties: true
...
    KnowledgeGovernanceJobCreateResponse:
      type: object
      required: [job_id, status, queued_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
        queued_at:
          type: string
          format: date-time
...
paths:
...
  /internal/memory/query:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryQueryRequest"
      responses:
        "200":
          description: Memory query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryQueryResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/memory/candidates:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryCandidateRequest"
      responses:
        "200":
          description: Candidate ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryCandidateResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/memory/retrieve:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryRetrieveRequest"
      responses:
        "200":
          description: Retrieval bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRetrieveResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/rules/gate/check:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleGateCheckRequest"
      responses:
        "200":
          description: Rule gate decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleGateCheckResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/memory/governance/run:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GovernanceRunRequest"
      responses:
        "200":
          description: Governance completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceRunResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/governance/jobs:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeGovernanceJobCreateRequest"
      responses:
        "200":
          description: Governance job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGovernanceJobCreateResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/governance/jobs/{jobId}:
    get:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Governance job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGovernanceJobResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/governance/run:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeGovernanceRunRequest"
      responses:
        "200":
          description: Knowledge governance completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGovernanceRunResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/documents/ingest:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeDocumentIngestRequest"
      responses:
        "200":
          description: Knowledge document ingested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocumentIngestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/documents:
    get:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Knowledge documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeDocumentListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/review-queue:
    get:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Review queue items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeReviewQueueResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/review-queue/{reviewQueueId}/actions:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: reviewQueueId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeReviewActionRequest"
      responses:
        "200":
          description: Review action applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeReviewActionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/context-bundles/{bundleId}:
    get:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Context bundle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeContextBundleResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/retrieve:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeRetrieveRequest"
      responses:
        "200":
          description: Knowledge retrieval bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeRetrieveResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/knowledge/graph/entities:
    get:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Knowledge graph entities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeGraphListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...